重构代码并提高代码质量(Excel)
- 删除了多个类中的冗余代码和注释
- 优化了部分方法的实现，提高了代码可读性和性能
- 统一了ApiResult的泛型使用，提高了类型安全性
- 修复了一些潜在的空指针异常问题

+++ geelato-web-platform/src/main/java/cn/geelato/web/platform/m/excel/entity/WordWaterMarkMeta.java
@@ -14,7 +14,7 @@ import java.util.regex.Pattern;
 
 /**
  * @author diabl
- * @description: 水印
+ * 水印
  */
 @Getter
 @Setter
@@ -72,7 +72,7 @@ public class WordWaterMarkMeta {
         if (Strings.isBlank(this.fontColor)) {
             setFontColor(WordWaterMarkMeta.FONT_COLOR);
         }
-        if (this.fontSize == null || this.fontSize.doubleValue() <= 0) {
+        if (this.fontSize == null || this.fontSize <= 0) {
             setFontSize(WordWaterMarkMeta.FONT_SIZE);
         }
     }
